+++ internal/interfaces/repositories.go
@@ -7,14 +7,14 @@ import (
 )
 
 type UsersRepository interface {
-	GetUserByID(id int) (*entities.User, error)
+	GetUserByID(id uint64) (*entities.User, error)
 	GetAllUsers() ([]*entities.User, error)
 	GetUserByEmail(email string) (*entities.User, error)
 }
 
 type AuthRepository interface {
-	RegisterUser(userData entities.RegisterUserDTO) (userID int, err error)
-	CreateRefreshToken(userID int, refreshToken string, ttl time.Duration) (refreshTokenID int, err error)
-	GetRefreshTokenByUserID(userID int) (*entities.RefreshToken, error)
+	RegisterUser(userData entities.RegisterUserDTO) (userID uint64, err error)
+	CreateRefreshToken(userID uint64, refreshToken string, ttl time.Duration) (refreshTokenID uint64, err error)
+	GetRefreshTokenByUserID(userID uint64) (*entities.RefreshToken, error)
 	ExpireRefreshToken(refreshToken string) error
 }
