+++ internal/interfaces/usecases.go
@@ -5,9 +5,9 @@ import (
 )
 
 type UseCases interface {
-	GetUserByID(id int) (*entities.User, error)
+	GetUserByID(id uint64) (*entities.User, error)
 	GetAllUsers() ([]*entities.User, error)
-	RegisterUser(userData entities.RegisterUserDTO) (userID int, err error)
+	RegisterUser(userData entities.RegisterUserDTO) (userID uint64, err error)
 	LoginUser(userData entities.LoginUserDTO) (*entities.TokensDTO, error)
 	GetMe(accessToken string) (*entities.User, error)
 	RefreshTokens(refreshTokensData entities.TokensDTO) (*entities.TokensDTO, error)
